feature/#43 :SchoolInfoResponseDto JAXB 디코더 적용
